5th commit.
Added custom /error mapping, and a home button on some of the pages. Also packaged the project in a .jar file and created a .bat file that, when executed, will run the .jar file from command console. CONCLUSION By no means I'd say this is finished, but it does fulfill the given requirements.